NU-2159 refactor
