+++ designer/client/src/components/graph/node-modal/editors/expression/CustomCompleterAceEditor.tsx
@@ -1,26 +1,21 @@
 import { cx } from "@emotion/css";
 import { Box, Fade, LinearProgress, styled } from "@mui/material";
 import { isEmpty } from "lodash";
-import type { ReactNode} from "react";
-import { useMemo } from "react";
+import type { ReactNode } from "react";
 import React, { useCallback, useState } from "react";
 import type ReactAce from "react-ace/lib/ace";
 import { useSelector } from "react-redux";
 
-import { useUserSettings } from "../../../../../common/userSettings";
 import { getUserSettings } from "../../../../../reducers/selectors/userSettings";
-import type { UserSettings } from "../../../../../reducers/userSettings";
 import ValidationLabels from "../../../../modals/ValidationLabels";
 import { nodeInputCss } from "../../../../NodeInput";
 import { nodeInput, nodeInputWithError, nodeValue, rowAceEditor } from "../../NodeDetailsContent/NodeTableStyled";
 import type { FieldError } from "../Validators";
 import AceEditor from "./AceWithSettings";
 import type { AceWrapperInputProps } from "./AceWrapper";
-import { useAceEditorMessages } from "./AceWrapper";
 import type { CustomAceEditorCompleter } from "./CustomAceEditorCompleter";
 import type { ExpressionLang } from "./types";
-
-
+import { useAceEditorRangeMessages } from "./useAceEditorRangeMessages";
 
 type InputProps = AceWrapperInputProps & {
     language: ExpressionLang | string;
@@ -53,7 +48,7 @@ export function CustomCompleterAceEditor(props: CustomCompleterAceEditorProps): 
 
     const onChange = useCallback((value: string) => onValueChange(value), [onValueChange]);
     const editorFocus = useCallback((editorFocused: boolean) => () => setEditorFocused(editorFocused), []);
-    const { annotations, markers } = useAceEditorMessages(fieldErrors);
+    const { annotations, markers } = useAceEditorRangeMessages(fieldErrors);
 
     return (
         <Box className={cx(nodeValue, className)} sx={{ width: "100%" }}>
